feat(theme): nav links on the left (close: 73)
