+++ packages/vuepress-theme-vt/components/Navbar.vue
@@ -19,32 +19,12 @@
         >
       </RouterLink>
 
-      <div
-        class="links"
-        :style="
-          linksWrapMaxWidth
-            ? {
-                'max-width': linksWrapMaxWidth + 'px',
-              }
-            : {}
-        "
-      >
-        <AlgoliaSearchBox v-if="isAlgoliaSearch" :options="algolia" />
-        <SearchBox
-          v-else-if="
-            $site.themeConfig.search !== false &&
-            $page.frontmatter.search !== false
-          "
-        />
-        <NavLinks class="can-hide" />
-      </div>
+      <NavLinks class="can-hide" />
     </div>
   </header>
 </template>
 
 <script>
-import AlgoliaSearchBox from "@AlgoliaSearchBox";
-import SearchBox from "@SearchBox";
 import SidebarButton from "@theme/components/SidebarButton.vue";
 import NavLinks from "@theme/components/NavLinks.vue";
 
@@ -54,54 +34,8 @@ export default {
   components: {
     SidebarButton,
     NavLinks,
-    SearchBox,
-    AlgoliaSearchBox,
-  },
-
-  data() {
-    return {
-      linksWrapMaxWidth: null,
-    };
-  },
-
-  computed: {
-    algolia() {
-      return (
-        this.$themeLocaleConfig.algolia || this.$site.themeConfig.algolia || {}
-      );
-    },
-
-    isAlgoliaSearch() {
-      return this.algolia && this.algolia.apiKey && this.algolia.indexName;
-    },
-  },
-
-  mounted() {
-    const MOBILE_DESKTOP_BREAKPOINT = 719; // refer to config.styl
-    const NAVBAR_VERTICAL_PADDING =
-      parseInt(css(this.$el, "paddingLeft")) +
-      parseInt(css(this.$el, "paddingRight"));
-    const handleLinksWrapWidth = () => {
-      if (document.documentElement.clientWidth < MOBILE_DESKTOP_BREAKPOINT) {
-        this.linksWrapMaxWidth = null;
-      } else {
-        this.linksWrapMaxWidth =
-          this.$el.offsetWidth -
-          NAVBAR_VERTICAL_PADDING -
-          ((this.$refs.siteName && this.$refs.siteName.offsetWidth) || 0);
-      }
-    };
-    handleLinksWrapWidth();
-    window.addEventListener("resize", handleLinksWrapWidth, false);
   },
 };
-
-function css(el, property) {
-  // NOTE: Known bug, will return 'auto' if style value is 'auto'
-  const win = el.ownerDocument.defaultView;
-  // null means not to return pseudo styles
-  return win.getComputedStyle(el, null)[property];
-}
 </script>
 
 <style lang="stylus">
@@ -151,22 +85,6 @@ function css(el, property) {
     position: relative;
   }
 
-  .links {
-    height: 100%;
-    padding-left: 1.5rem;
-    box-sizing: border-box;
-    white-space: nowrap;
-    font-size: 13px;
-    display: flex;
-    align-items: center;
-    justify-content: flex-end;
-    flex: 1;
-
-    .search-box {
-      flex: 0 0 auto;
-      vertical-align: top;
-    }
-  }
 }
 
 @media (max-width: $MQMobile) {
@@ -177,10 +95,6 @@ function css(el, property) {
 
     .can-hide {
       display: none;
-    }
-
-    .links {
-      padding-left: 1.5rem;
     }
 
     .site-name {
