+++ packages/vuepress-theme-vt/styles/config.styl
@@ -1,5 +1,5 @@
 $contentClass = '.theme-default-content';
-$navbarHeight = 3rem
+$navbarHeight = 3.6rem
 
 /**
  * CSS Varible cannot be used at media query, so we used stylus varaible here.
