+++ packages/vuepress-theme-vt/styles/index.styl
@@ -7,7 +7,6 @@
 @require './wrapper';
 @require './toc';
 @require './header';
-@require './mobile';
 @require './home';
 @require './vp-doc';
 
@@ -235,3 +234,5 @@ table {
     }
   }
 }
+
+@require './mobile';
